+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Event follow up survey preview and entry point [#3439](https://github.com/rokwire/illinois-app/issues/3439).
 ### Deleted
 - Removed Events2 attributes image background [#3442](https://github.com/rokwire/illinois-app/issues/3442).
+- Removed the references to the old events in Home and Browse panels [#3450](https://github.com/rokwire/illinois-app/issues/3450).
 ### Changed
 - Updated Event2 onboarding UI [#3445](https://github.com/rokwire/illinois-app/issues/3445).
 - Use extended error messaging for Events2 content [#3447](https://github.com/rokwire/illinois-app/issues/3447).